Reorganise impedance
